fixup! Rework error handling
