+++ crates/apollo-compiler/src/executable/filtering.rs
@@ -6,9 +6,9 @@ use crate::executable::InlineFragment;
 use crate::executable::Operation;
 use crate::executable::Selection;
 use crate::executable::SelectionSet;
-use crate::execution::SuspectedValidationBug;
 use crate::schema;
 use crate::schema::Name;
+use crate::validation::SuspectedValidationBug;
 use crate::validation::Valid;
 use crate::ExecutableDocument;
 use crate::Node;
