+++ crates/apollo-compiler/src/execution/introspection.rs
@@ -7,9 +7,9 @@ use crate::execution::engine::ExecutionMode;
 use crate::execution::resolver::ResolvedValue;
 use crate::execution::JsonMap;
 use crate::execution::Response;
-use crate::execution::SuspectedValidationBug;
 use crate::schema;
 use crate::schema::Name;
+use crate::validation::SuspectedValidationBug;
 use crate::validation::Valid;
 use crate::ExecutableDocument;
 use crate::Node;
