+++ crates/apollo-compiler/src/execution/mod.rs
@@ -13,13 +13,13 @@ mod response;
 mod result_coercion;
 
 pub use self::input_coercion::coerce_variable_values;
+pub use self::input_coercion::InputCoercionError;
 pub use self::introspection::SchemaIntrospection;
 pub use self::response::GraphQLError;
 pub use self::response::GraphQLLocation;
 pub use self::response::PathElement;
 pub use self::response::Response;
 pub use self::response::ResponseData;
-pub use self::response::SuspectedValidationBug;
 /// Re-export of the version of the `serde_json_bytes` crate used for [`JsonValue`] and [`JsonMap`]
 pub use serde_json_bytes;
